refactor(ucmc-web): extract features/audit/ viewer
The audit-log viewer at /audit is a user-facing admin feature: list
events, filter by action type / date, paginate. It belongs alongside
the other features under src/features/, not in src/server/audit/ next
to the recorder. Move the viewer's server fn shell + action + viewer
UI into a new features/audit/ following the standard
api/{queries,query-keys}.ts + components/ + server/ shape, and switch
the route from an inline `useQuery({ queryKey, queryFn })` to
`useQuery(auditEventsQueryOptions(...))`.

Recorder stays shared in src/server/audit/audit-log.server.ts — every
feature writes audit events, so it's cross-feature foundational. Tests
for the recorder stay in src/server/audit/__tests__; tests for the
viewer's listAuditEventsAction move into the feature alongside the
action it covers.

The route's AuditCard / ActorLabel / TargetLabel components move into
features/audit/components/, taking the USER_TARGETED_ACTIONS set with
them. AUDIT_ACTIONS is now exported from server/audit-fns.ts so the
filter Select doesn't have to maintain its own duplicate of the enum.
