+++ CLAUDE.md
@@ -74,11 +74,12 @@ This is the software system for the University of Cincinnati Mountaineering Club
 - **Build tool**: Vite 8 (config in `apps/web/vite.config.ts`)
 - **Styling**: Tailwind CSS v4 (via `@tailwindcss/vite`), shadcn components
 - **State/data**: TanStack Query, TanStack Router (with SSR query integration), TanStack Form, TanStack Table
-- **Source layout** (Bulletproof React–aligned): the app is organized into four features under `src/features/`:
+- **Source layout** (Bulletproof React–aligned): the app is organized into features under `src/features/`:
   - `features/auth/` — sign-in flow (magic link, WebAuthn passkey), session management, the user's own profile/avatar editing, and the `useAuth`/`view-mode`/`guards.ts` "auth public API" surface that other features depend on.
   - `features/members/` — registration approval queue, member directory, admin profile editing, role + permission management (RBAC).
   - `features/announcements/` — announcements feed with admin authoring + read-tracking.
   - `features/feedback/` — in-app bug / feature / general / question submissions with an admin triage view, optionally mirrored to GitHub Issues server-side.
+  - `features/audit/` — audit-log viewer at `/audit` (read-only feed of officer/admin actions). The recorder side that every feature calls — `recordAuditEvent`, `buildAuditEventStatement`, `buildBulkAuditEventStatement` — stays in `src/server/audit/audit-log.server.ts` because it's cross-feature foundational; only the read-side server fn + viewer UI live in this feature. No other feature imports from `features/audit/`.
     Each feature has its own `components/`, `api/` (query options + per-mutation `use-*.ts` hooks + `query-keys.ts`), and `server/` (server fns, action implementations, repositories, tests). Shared/foundational code lives outside features: `src/server/auth/` (session, principal, cookies, ids), `src/server/profile/` (zod schemas + constants), `src/components/` (UI primitives + `user-avatar.tsx` + `profile/` field components + `layouts/` + `markdown/` markdown viewer + `editor/` markdown editor), `src/lib/` (utilities), `src/hooks/`, `src/config/`. Routes (`src/routes/`) compose features but never the reverse. The `import/no-restricted-paths` rule in `apps/web/eslint.config.js` enforces these boundaries mechanically — features can't import each other (auth's `api/use-auth.ts`, `api/view-mode.tsx`, and `guards.ts` are the narrow exception), and shared code (`components/ui/`, `components/profile/`, `lib/`, `hooks/`, `config/`) can't import features. Tests are exempt because integration tests legitimately span features.
 - **API layer**: every feature follows the same query/mutation pattern in `features/<name>/api/`. `query-keys.ts` centralizes the keys. `queries.ts` exposes `*QueryOptions()` factories that callers spread into `useQuery(...)`. Each mutation has its own `use-*.ts` hook that wraps `useMutation` with a fixed cache-invalidation contract — the call site decides toast/navigate behavior via `mutate(input, { onSuccess, onError })`. Inline `useMutation` calls in routes/components were eliminated in Phase 6 of the Bulletproof refactor.
 - **Env**: `@t3-oss/env-core` + zod (`apps/web/src/config/env.ts`) for client (`VITE_*`) vars. Server/Worker vars (APP_BASE_URL, WEBAUTHN_RP_ID, WEBAUTHN_RP_NAME, RESEND_FROM, RESEND_FROM_NAME, RESEND_API_KEY, SESSION_SECRET, MAILPIT_URL) are not in t3-env — they reach request handlers via the Worker `env` binding (`apps/web/src/server/cloudflare-env.ts`). Local dev loads them from `apps/web/.env.local` (copy `.env.example`, gitignored) through wrangler v4's standard `.env` precedence (`.env.<mode>.local` > `.env.local` > `.env.<mode>` > `.env`); `.dev.vars` is no longer used. The devcontainer sets `CLOUDFLARE_INCLUDE_PROCESS_ENV=true` in `docker-compose.yml`, so host-exported env vars (e.g. `SESSION_SECRET`, `RESEND_API_KEY` from the developer's shell) win over `.env.local`. Deployed envs get these same vars from Pulumi (`--var` flags) and GitHub secrets (`wrangler secret put`) via `web-deploy.yml`. Email sending follows a three-tier fallback in `apps/web/src/server/email/resend.ts`: (1) Resend API when `RESEND_API_KEY` is set, (2) Mailpit HTTP send API when `MAILPIT_URL` is set (dev sidecar — emails appear at http://localhost:8025), (3) Worker console log when neither is set.
